third_party/blink/web_tests/external/wpt/css/css-sizing/intrinsic-size-fallback-video-expected.txt

This is a testharness.js-based test.
[FAIL] .wrapper 2
  assert_equals: \n<div class="wrapper" data-expected-client-width="200">\n  <video style="height: 100px"></video>\n</div>\nclientWidth expected 200 but got 300
[FAIL] .wrapper 3
  assert_equals: \n<div class="wrapper" data-expected-client-width="100">\n  <video style="max-height: 50px"></video>\n</div>\nclientWidth expected 100 but got 300
[FAIL] .wrapper 4
  assert_equals: \n<div class="wrapper" data-expected-client-width="400">\n  <video style="min-height: 200px"></video>\n</div>\nclientWidth expected 400 but got 300
Harness: the test ran to completion.
